+++ src/Schema/Values/ContextModule.ts
@@ -6,6 +6,7 @@
  */
 export enum ContextModule {
   aboutTheWork = "aboutTheWork",
+  aboutThisAuction = "aboutThisAuction",
   articleArtist = "articleArtist",
   artistCard = "artistCard",
   artistHeader = "artistHeader",
@@ -22,6 +23,7 @@ export enum ContextModule {
   associatedViewingRoom = "associatedViewingRoom",
   auctionSidebar = "auctionSidebar",
   auctionCard = "auctionCard",
+  auctionHome = "auctionHome",
   auctionRail = "auctionRail",
   auctionResults = "auctionResults",
   auctionsInfo = "auctionsInfo",
@@ -51,6 +53,7 @@ export enum ContextModule {
   header = "header",
   intextTooltip = "intextTooltip",
   latestViewingRoomsRail = "latestViewingRoomsRail",
+  liveAuctionRoom = "liveAuctionRoom",
   liveAuctionsRail = "liveAuctionsRail",
   mainCarousel = "mainCarousel",
   minimalCTABanner = "minimalCTABanner",
@@ -70,6 +73,7 @@ export enum ContextModule {
   presentingFair = "presentingFair",
   presentingPartner = "presentingPartner",
   pushNotifications = "pushNotifications",
+  priceEstimate = "priceEstimate",
   promoSpace = "promoSpace",
   recentlySavedRail = "recentlySavedRail",
   recentlyViewedRail = "recentlyViewedRail",
@@ -112,6 +116,7 @@ export type AuthContextModule =
   | ContextModule.associatedViewingRoom
   | ContextModule.auctionSidebar
   | ContextModule.auctionRail
+  | ContextModule.auctionHome
   | ContextModule.auctionResults
   | ContextModule.auctionsInfo
   | ContextModule.bannerPopUp
@@ -129,6 +134,7 @@ export type AuthContextModule =
   | ContextModule.geneHeader
   | ContextModule.header
   | ContextModule.intextTooltip
+  | ContextModule.liveAuctionRoom
   | ContextModule.liveAuctionsRail
   | ContextModule.mainCarousel
   | ContextModule.minimalCTABanner
@@ -142,6 +148,7 @@ export type AuthContextModule =
   | ContextModule.popUpModal
   | ContextModule.presentingFair
   | ContextModule.presentingPartner
+  | ContextModule.priceEstimate
   | ContextModule.recentlyViewedRail
   | ContextModule.recommendedArtistsRail
   | ContextModule.relatedArtistsRail